deleted files

+++ app/Http/Controllers/FileController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\DestroyFileRequest;
 use App\Http\Requests\StoreFileRequest;
 use App\Http\Requests\StoreFolderRequest;
 use App\Http\Resources\FileResource;
@@ -111,7 +112,7 @@ class FileController extends Controller
         }
     }
 
-    public function destroy(Request $request)
+    public function destroy(DestroyFileRequest $request)
     {
         $data = $request->validated();
         $parent = $request->parent;
@@ -125,7 +126,10 @@ class FileController extends Controller
         } else {
             foreach ($data['ids'] as $id) {
                 $file = File::find($id);
-                $file->delete();
+                if($file){
+                    $file->delete();
+                }
+    
             }
         }
